chore: manually generate ion-icon standalone proxy
